+++ Forms/AppsForm.vb
@@ -223,14 +223,7 @@ Public Class AppsForm
         ScaleClientSize(41, 27)
         tv.ItemHeight = CInt(FontHeight * 1.5)
 
-        Dim plugins = Package.Items.Values.OfType(Of PluginPackage)
-        Dim x64 = Package.Items.Values.Where(Function(arg) Not arg.Version Is Nothing AndAlso Not arg.Version.Contains("x86")).Count
-        Dim avs = plugins.Where(Function(arg) Not arg.AviSynthFilterNames.NothingOrEmpty).Count
-        Dim vs = plugins.Where(Function(arg) Not arg.VapourSynthFilterNames.NothingOrEmpty).Count
-        Dim exe = Package.Items.Values.Where(Function(arg) arg.Filename.Ext = "exe" AndAlso Not arg.Version Is Nothing AndAlso arg.Version.Contains("x64")).Count
-        Dim dll = Package.Items.Values.Where(Function(arg) arg.Filename.Ext = "dll" AndAlso Not TypeOf arg Is PluginPackage).Count
-
-        Text = $"{x64} x64 packages   {avs} AVS x64 plugins   {exe} x64 tools   {vs} VS x64 plugins   {dll} x64 libraries"
+        Text = $"{Package.Items.Count} tools"
         SearchTextBox_TextChanged()
 
         tv.Scrollable = True
@@ -395,7 +388,7 @@ Public Class AppsForm
             Case Keys.F1
                 tsbHelp.PerformClick()
             Case Keys.F10
-                Dim fp = "D:\Projekte\VS\VB\StaxRip\md\test-build.md"
+                Dim fp = "D:\Projekte\VS\VB\StaxRip\md\changelog.md"
                 If File.Exists(fp) Then g.StartProcess(fp)
             Case Keys.F11
                 tsbPath.PerformClick()